requirements.txt
Django==3.0.5
manage.py==0.2.10
requests==2.22.0
djangorestframework==3.11.1
django-filter==2.3.0
